Move browser client certificate logic into CmuxBrowser
