--- Sources/Panels/BrowserClientCertificateCredentialCandidate.swift
@@ -1,17 +1,0 @@
-import Foundation
-
-struct BrowserClientCertificateCredentialCandidate {
-    let title: String?
-    let subtitle: String?
-    let credential: URLCredential
-
-    init(
-        title: String? = nil,
-        subtitle: String? = nil,
-        credential: URLCredential
-    ) {
-        self.title = title
-        self.subtitle = subtitle
-        self.credential = credential
-    }
-}
